nasty circular dependency
