+++ src/components/Game/Game.css
@@ -1,0 +1,4 @@
+.game {
+  display: flex;
+  justify-content: center;
+}
